Example1
This example demonstrates creating a toString method for a particular class the contains private variables, to be used to format and control how the value contents of an object of the class are displayed. Using String.format or StringBuilder is much more efficient rather than concatenating multiple strings to accomplish the same thing.